Move content extraction out of utils
