+++ tests/sources/test_sharepoint_online.py
@@ -2447,7 +2447,7 @@ class TestSharepointOnlineDataSource:
             )
 
     @pytest.mark.asyncio
-    @patch("connectors.utils.ExtractionService._check_configured", lambda *_: True)
+    @patch("connectors.content_extraction.ContentExtraction._check_configured", lambda *_: True)
     async def test_get_attachment_with_text_extraction_enabled_adds_body(
         self, patch_sharepoint_client
     ):
@@ -2455,9 +2455,9 @@ class TestSharepointOnlineDataSource:
         message = "This is the text content of drive item"
 
         with patch(
-            "connectors.utils.ExtractionService.extract_text", return_value=message
+            "connectors.content_extraction.ContentExtraction.extract_text", return_value=message
         ) as extraction_service_mock, patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
+            "connectors.content_extraction.ContentExtraction.get_extraction_config",
             return_value={"host": "http://localhost:8090"},
         ):
 
@@ -2475,7 +2475,7 @@ class TestSharepointOnlineDataSource:
                 assert "_attachment" not in download_result
 
     @pytest.mark.asyncio
-    @patch("connectors.utils.ExtractionService._check_configured", lambda *_: False)
+    @patch("connectors.content_extraction.ContentExtraction._check_configured", lambda *_: False)
     async def test_get_attachment_with_text_extraction_enabled_but_not_configured_adds_empty_string(
         self, patch_sharepoint_client
     ):
@@ -2483,9 +2483,9 @@ class TestSharepointOnlineDataSource:
         message = "This is the text content of drive item"
 
         with patch(
-            "connectors.utils.ExtractionService.extract_text", return_value=message
+            "connectors.content_extraction.ContentExtraction.extract_text", return_value=message
         ) as extraction_service_mock, patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
+            "connectors.content_extraction.ContentExtraction.get_extraction_config",
             return_value={"host": "http://localhost:8090"},
         ):
 
@@ -2506,7 +2506,7 @@ class TestSharepointOnlineDataSource:
     @pytest.mark.parametrize(
         "filesize, expect_download", [(15, True), (10485761, False)]
     )
-    @patch("connectors.utils.ExtractionService._check_configured", lambda *_: True)
+    @patch("connectors.content_extraction.ContentExtraction._check_configured", lambda *_: True)
     async def test_get_drive_item_content(
         self, patch_sharepoint_client, filesize, expect_download
     ):
@@ -2536,7 +2536,7 @@ class TestSharepointOnlineDataSource:
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize("filesize", [(15), (10485761)])
-    @patch("connectors.utils.ExtractionService._check_configured", lambda *_: True)
+    @patch("connectors.content_extraction.ContentExtraction._check_configured", lambda *_: True)
     async def test_get_content_with_text_extraction_enabled_adds_body(
         self, patch_sharepoint_client, filesize
     ):
@@ -2550,9 +2550,9 @@ class TestSharepointOnlineDataSource:
         message = "This is the text content of drive item"
 
         with patch(
-            "connectors.utils.ExtractionService.extract_text", return_value=message
+            "connectors.content_extraction.ContentExtraction.extract_text", return_value=message
         ) as extraction_service_mock, patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
+            "connectors.content_extraction.ContentExtraction.get_extraction_config",
             return_value={"host": "http://localhost:8090"},
         ):
 
@@ -2571,7 +2571,7 @@ class TestSharepointOnlineDataSource:
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize("filesize", [(15), (10485761)])
-    @patch("connectors.utils.ExtractionService._check_configured", lambda *_: False)
+    @patch("connectors.content_extraction.ContentExtraction._check_configured", lambda *_: False)
     async def test_get_content_with_text_extraction_enabled_but_not_configured_adds_empty_string(
         self, patch_sharepoint_client, filesize
     ):
@@ -2585,9 +2585,9 @@ class TestSharepointOnlineDataSource:
         message = "This is the text content of drive item"
 
         with patch(
-            "connectors.utils.ExtractionService.extract_text", return_value=message
+            "connectors.content_extraction.ContentExtraction.extract_text", return_value=message
         ) as extraction_service_mock, patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
+            "connectors.content_extraction.ContentExtraction.get_extraction_config",
             return_value={"host": "http://localhost:8090"},
         ):
