+++ tests/test_utils.py
@@ -16,18 +16,15 @@ import tempfile
 import time
 import timeit
 from datetime import datetime
-from unittest.mock import Mock, mock_open, patch
+from unittest.mock import Mock, patch
 
 import pytest
-import pytest_asyncio
-from aioresponses import aioresponses
 from freezegun import freeze_time
 from pympler import asizeof
 
 from connectors import utils
 from connectors.utils import (
     ConcurrentTasks,
-    ExtractionService,
     InvalidIndexNameError,
     MemQueue,
     RetryStrategy,
@@ -776,145 +773,6 @@ def test_iterable_batches_generator(iterable, batch_size_, expected_batches):
         actual_batches.append(batch)
 
     assert actual_batches == expected_batches
-
-
-class TestExtractionService:
-    @pytest_asyncio.fixture
-    async def mock_responses(self):
-        with aioresponses() as m:
-            yield m
-
-    @pytest.mark.parametrize(
-        "mock_config, expected_result",
-        [
-            (
-                {
-                    "extraction_service": {
-                        "host": "http://localhost:8090",
-                    }
-                },
-                True,
-            ),
-            ({"something_else": "???"}, False),
-            ({"extraction_service": {"not_a_host": "!!!m"}}, False),
-        ],
-    )
-    def test_check_configured(self, mock_config, expected_result):
-        with patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
-            return_value=mock_config.get("extraction_service", None),
-        ):
-            extraction_service = ExtractionService()
-            assert extraction_service._check_configured() is expected_result
-
-    @pytest.mark.asyncio
-    async def test_extract_text(self, mock_responses, patch_logger):
-        filepath = "tmp/notreal.txt"
-        url = "http://localhost:8090/extract_text/"
-        payload = {"extracted_text": "I've been extracted!"}
-
-        with patch("builtins.open", mock_open(read_data=b"data")), patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
-            return_value={"host": "http://localhost:8090"},
-        ):
-            mock_responses.put(url, status=200, payload=payload)
-
-            extraction_service = ExtractionService()
-            extraction_service._begin_session()
-
-            response = await extraction_service.extract_text(filepath, "notreal.txt")
-            await extraction_service._end_session()
-
-            assert response == "I've been extracted!"
-            patch_logger.assert_present(
-                "Text extraction is successful for 'notreal.txt'."
-            )
-
-    @pytest.mark.asyncio
-    async def test_extract_text_with_file_pointer(self, mock_responses, patch_logger):
-        filepath = "/tmp/notreal.txt"
-        url = "http://localhost:8090/extract_text/?local_file_path=/tmp/notreal.txt"
-        payload = {"extracted_text": "I've been extracted from a local file!"}
-
-        with patch("builtins.open", mock_open(read_data=b"data")), patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
-            return_value={
-                "host": "http://localhost:8090",
-                "use_file_pointers": True,
-                "shared_volume_dir": "/tmp",
-            },
-        ):
-            mock_responses.put(url, status=200, payload=payload)
-
-            extraction_service = ExtractionService()
-            extraction_service._begin_session()
-
-            response = await extraction_service.extract_text(filepath, "notreal.txt")
-            await extraction_service._end_session()
-
-            assert response == "I've been extracted from a local file!"
-            patch_logger.assert_present(
-                "Text extraction is successful for 'notreal.txt'."
-            )
-
-    @pytest.mark.asyncio
-    async def test_extract_text_when_response_isnt_200_logs_warning(
-        self, mock_responses, patch_logger
-    ):
-        filepath = "tmp/notreal.txt"
-        url = "http://localhost:8090/extract_text/"
-
-        with patch("builtins.open", mock_open(read_data=b"data")), patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
-            return_value={"host": "http://localhost:8090"},
-        ):
-            mock_responses.put(
-                url,
-                status=422,
-                payload={
-                    "error": "Unprocessable Entity",
-                    "message": "Could not process file.",
-                },
-            )
-
-            extraction_service = ExtractionService()
-            extraction_service._begin_session()
-
-            response = await extraction_service.extract_text(filepath, "notreal.txt")
-            await extraction_service._end_session()
-            assert response == ""
-
-            patch_logger.assert_present(
-                "Extraction service could not parse `notreal.txt'. Status: [422]; Unprocessable Entity: Could not process file."
-            )
-
-    @pytest.mark.asyncio
-    async def test_extract_text_when_response_is_200_with_error_logs_warning(
-        self, mock_responses, patch_logger
-    ):
-        filepath = "tmp/notreal.txt"
-        url = "http://localhost:8090/extract_text/"
-
-        with patch("builtins.open", mock_open(read_data=b"data")), patch(
-            "connectors.utils.ExtractionService.get_extraction_config",
-            return_value={"host": "http://localhost:8090"},
-        ):
-            mock_responses.put(
-                url,
-                status=200,
-                payload={"error": "oh no!", "message": "I'm all messed up..."},
-            )
-
-            extraction_service = ExtractionService()
-            extraction_service._begin_session()
-
-            response = await extraction_service.extract_text(filepath, "notreal.txt")
-            await extraction_service._end_session()
-            assert response == ""
-
-            patch_logger.assert_present(
-                "Extraction service could not parse `notreal.txt'. Status: [200]; oh no!: I'm all messed up..."
-            )
 
 
 @pytest.mark.parametrize(
